bug fix in skill

+++ src/pilots/Skill.hpp
@@ -34,8 +34,8 @@ class Skill
 		int GetExpirienceNextLevel() const 	{ return expirience_nextlevel; };
 		int GetLevel() const { return level; };
 
-		int GetAttackNormalized()   const { return attack  * SKILL::ATTACK_NORMALIZED_RATE; };   
-		int GetDefenceNormalized()  const { return defence * SKILL::DEFENCE_NORMALIZED_RATE; }; 
+		float GetAttackNormalized()   const { return attack  * SKILL::ATTACK_NORMALIZED_RATE; };   
+		float GetDefenceNormalized()  const { return defence * SKILL::DEFENCE_NORMALIZED_RATE; }; 
 		
 		int GetAttack()   const { return attack; };   
 		int GetDefence()  const { return defence; };
